[FLOW] Use positions at same time as velocity for flow field collection
Positions and velocities are staggered when using the leap-frog
integrator. For this integrator we now project the positions one
half-step backwards to get them at the same times as the velocity. We
then use this positions to get the bin for the current atom.

This fix was suggested and originally implemented by Michele Pellegrino.