[flow] Fix bug where first flow field output averages with wrong count
At step 0 we collect but do not output, and at some later step N we also
collect but this time output. Since flow field data is reset after
output, this means that the first flow field has 1 extra sample from
step 0 compared to all other flow fields. However, this was not
accounted for when averaging the data.

We fix this by keeping a correct count of the number of samples since
the last output.