[flow] Exit mdrun with error if `-flow` is supplied without collection
This is to catch old run scripts from executing without enabling
collection in the .mdp file, thinking that the `-flow` flag is what
enables the collection. We could raise a warning and write it to the
log, but I think it's better to forcefully notify the user that their
expected collection is not being performed.

Additionally, we cannot simply enable it when that flag is supplied,
since variable verification (grid shape, etc.) is done by the
pre-processor.