[flow] Fix MPI communication of flow field data
It was previously written with the assumption that the transmitted data
array contained raw doubles. Since we changed this to contain
std::arrays of doubles, it was no longer transmitting all of the data:
only as many as the number of bins in the system, not knowing that each
bin contains (currently) five doubles.

This has been fixed, and a release assert has been added to ensure that
all bins are layed out contiguously in memory (which is needed for the
transmission to work).